// src/generated/prisma/models/JinisChara.ts
export type JinisCharaWhereInput = {
  AND?: Prisma.JinisCharaWhereInput | Prisma.JinisCharaWhereInput[]
  OR?: Prisma.JinisCharaWhereInput[]
  NOT?: Prisma.JinisCharaWhereInput | Prisma.JinisCharaWhereInput[]
  id?: Prisma.StringFilter<"JinisChara"> | string
  slNo?: Prisma.IntFilter<"JinisChara"> | number
  name?: Prisma.StringFilter<"JinisChara"> | string
  fatherName?: Prisma.StringFilter<"JinisChara"> | string
  phoneNo?: Prisma.StringFilter<"JinisChara"> | string
  credit?: Prisma.IntFilter<"JinisChara"> | number
  percentage?: Prisma.FloatFilter<"JinisChara"> | number
  description?: Prisma.StringNullableFilter<"JinisChara"> | string | null
  date?: Prisma.DateTimeFilter<"JinisChara"> | Date | string
  active?: Prisma.BoolFilter<"JinisChara"> | boolean
  settledAt?: Prisma.DateTimeNullableFilter<"JinisChara"> | Date | string | null
  createdById?: Prisma.StringFilter<"JinisChara"> | string
  createdAt?: Prisma.DateTimeFilter<"JinisChara"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"JinisChara"> | Date | string
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
  interests?: Prisma.InterestListRelationFilter
}